+++ build.gradle
@@ -33,8 +33,8 @@ application {
 }
 
 shadowJar {
-    archiveBaseName.set("duke")
-    archiveClassifier.set("")
+    archiveBaseName = 'BusynessManager'
+    archiveVersion = '1.0'
 }
 
 checkstyle {
