+++ src/main/java/busynessmanager/BusynessManager.java
@@ -1,41 +1,43 @@
 package busynessmanager;
 
+import busynessmanager.managers.InventoryManager;
+import busynessmanager.managers.SalesManager;
+import busynessmanager.managers.SearchManager;
+import busynessmanager.revenue.RevenueCalculator;
 import busynessmanager.parser.CommandParser;
+import busynessmanager.ui.UI;
+
+import static busynessmanager.constants.Constants.BM_UPPERCASE_REGEX;
+import static busynessmanager.constants.Constants.BM_BUSINESSTYPE_FNB;
+import static busynessmanager.constants.Constants.BM_BUSINESSTYPE_RETAIL;
+import static busynessmanager.constants.Constants.BM_WELCOME_MESSAGE;
+import static busynessmanager.constants.Constants.BM_NO_INPUT_ERROR_MESSAGE;
+import static busynessmanager.constants.Constants.BM_ENTER_BUSINESS_ID_MESSAGE;
+import static busynessmanager.constants.Constants.BM_ENTER_PASSWORD_MESSAGE;
+import static busynessmanager.constants.Constants.BM_ENTER_PASSWORD_MESSAGE_2;
+import static busynessmanager.constants.Constants.BM_SUCCESSFUL_LOGIN_MESSAGE;
+import static busynessmanager.constants.Constants.BM_INVALID_CREDENTIALS_MESSAGE;
+import static busynessmanager.constants.Constants.BM_FIRST_SETUP_MESSAGE;
+import static busynessmanager.constants.Constants.BM_ENTER_NAME_MESSAGE;
+import static busynessmanager.constants.Constants.BM_ENTER_BUSINESS_TYPE_MESSAGE;
+import static busynessmanager.constants.Constants.BM_INVALID_BUSINESSTYPE_ERROR_MESSAGE;
+import static busynessmanager.constants.Constants.BM_SETUP_COMPLETE_MESSAGE;
+import static busynessmanager.constants.Constants.BM_READY_MESSAGE;
+import static busynessmanager.constants.Constants.BM_WAITING_INPUT_MESSAGE;
+import static busynessmanager.constants.Constants.BM_EXIT_KEYWORD;
+import static busynessmanager.constants.Constants.BM_EXIT_MESSAGE;
+
 import java.util.Scanner;
 import java.util.HashMap;
 
-import busynessmanager.managers.InventoryManager;
-import busynessmanager.revenue.RevenueCalculator;
-import busynessmanager.managers.SalesManager;
-import busynessmanager.managers.SearchManager;
-import busynessmanager.UI_Constants.UI;
-
-import static busynessmanager.UI_Constants.Constants.BM_UPPERCASE_REGEX;
-import static busynessmanager.UI_Constants.Constants.BM_BUSINESSTYPE_FNB;
-import static busynessmanager.UI_Constants.Constants.BM_BUSINESSTYPE_RETAIL;
-import static busynessmanager.UI_Constants.Constants.BM_WELCOME_MESSAGE;
-import static busynessmanager.UI_Constants.Constants.BM_NO_INPUT_ERROR_MESSAGE;
-import static busynessmanager.UI_Constants.Constants.BM_ENTER_BUSINESS_ID_MESSAGE;
-import static busynessmanager.UI_Constants.Constants.BM_ENTER_PASSWORD_MESSAGE;
-import static busynessmanager.UI_Constants.Constants.BM_ENTER_PASSWORD_MESSAGE_2;
-import static busynessmanager.UI_Constants.Constants.BM_SUCCESSFUL_LOGIN_MESSAGE;
-import static busynessmanager.UI_Constants.Constants.BM_INVALID_CREDENTIALS_MESSAGE;
-import static busynessmanager.UI_Constants.Constants.BM_FIRST_SETUP_MESSAGE;
-import static busynessmanager.UI_Constants.Constants.BM_ENTER_NAME_MESSAGE;
-import static busynessmanager.UI_Constants.Constants.BM_ENTER_BUSINESS_TYPE_MESSAGE;
-import static busynessmanager.UI_Constants.Constants.BM_INVALID_BUSINESSTYPE_ERROR_MESSAGE;
-import static busynessmanager.UI_Constants.Constants.BM_SETUP_COMPLETE_MESSAGE;
-import static busynessmanager.UI_Constants.Constants.BM_READY_MESSAGE;
-import static busynessmanager.UI_Constants.Constants.BM_WAITING_INPUT_MESSAGE;
-import static busynessmanager.UI_Constants.Constants.BM_EXIT_KEYWORD;
-import static busynessmanager.UI_Constants.Constants.BM_EXIT_MESSAGE;
 
 /**
  * The main class for the Busyness Manager application.
  * Handles user authentication, business setup, and command execution.
  */
 public class BusynessManager {
-    private static final HashMap<String, String> credentials = new HashMap<>(); // Stores business ID & passwords
+    // Stores business ID & passwords
+    private static final HashMap<String, String> credentials = new HashMap<>();
 
     private enum BusinessType {
         FNB, RETAIL
@@ -47,6 +49,7 @@ public class BusynessManager {
     private BusinessType businessType;
 
     private final CommandParser commandParser;
+
 
     /**
      * Constructs a BusynessManager instance and initializes the necessary managers.
@@ -75,13 +78,11 @@ public class BusynessManager {
      */
     public void start() {
         Scanner scanner = new Scanner(System.in);
-        //System.out.println("Welcome to Busyness Manager!");
-        UI.printMessage(BM_WELCOME_MESSAGE);
 
-        //System.out.print("Enter Business ID: ");
-        UI.printMessage(BM_ENTER_BUSINESS_ID_MESSAGE);
+        UI.printMessage(BM_WELCOME_MESSAGE);
+        UI.printMessageWithoutNewline(BM_ENTER_BUSINESS_ID_MESSAGE);
+
         if (!scanner.hasNextLine()) {
-            //System.err.println("Error: No input detected. Exiting...");
             UI.printErrorMessage(BM_NO_INPUT_ERROR_MESSAGE);
             return;
         }
@@ -103,20 +104,19 @@ public class BusynessManager {
      * @param id The business ID entered by the user.
      */
     public void login(Scanner scanner, String id) {
-        //System.out.print("Enter Password: ");
-        UI.printMessage(BM_ENTER_PASSWORD_MESSAGE);
+        UI.printMessageWithoutNewline(BM_ENTER_PASSWORD_MESSAGE);
+
         if (!scanner.hasNextLine()) {
-            //System.err.println("Error: No input detected. Exiting...");
             UI.printErrorMessage(BM_NO_INPUT_ERROR_MESSAGE);
             return;
         }
+
         String password = scanner.nextLine().trim();
+
         if (validPassword(id, password)) {
-            //System.out.println("Login successful!");
             UI.printMessage(BM_SUCCESSFUL_LOGIN_MESSAGE);
             run(scanner);
         } else {
-            //System.out.println("Invalid credentials. Exiting.");
             UI.printMessage(BM_INVALID_CREDENTIALS_MESSAGE);
         }
     }
@@ -128,31 +128,26 @@ public class BusynessManager {
      * @param id The business ID entered by the user.
      */
     public void firstTimeSetup(Scanner scanner, String id) {
-        //System.out.println("First-time setup required.");
         UI.printMessage(BM_FIRST_SETUP_MESSAGE);
+        UI.printMessageWithoutNewline(BM_ENTER_NAME_MESSAGE);
 
-        //System.out.print("Enter Business Name: ");
-        UI.printMessage(BM_ENTER_NAME_MESSAGE);
         if (!scanner.hasNextLine()) {
-            //System.err.println("Error: No input detected. Exiting...");
             UI.printErrorMessage(BM_NO_INPUT_ERROR_MESSAGE);
             return;
         }
+
         businessName = scanner.nextLine().trim();
 
-        //System.out.print("Enter Business Password: ");
-        UI.printMessage(BM_ENTER_PASSWORD_MESSAGE_2);
+        UI.printMessageWithoutNewline(BM_ENTER_PASSWORD_MESSAGE_2);
         if (!scanner.hasNextLine()) {
-            //System.err.println("Error: No input detected. Exiting...");
             UI.printErrorMessage(BM_NO_INPUT_ERROR_MESSAGE);
             return;
         }
+
         businessPassword = scanner.nextLine().trim();
 
-        //System.out.print("Enter Business Type (FNB/RETAIL): ");
-        UI.printMessage(BM_ENTER_BUSINESS_TYPE_MESSAGE);
+        UI.printMessageWithoutNewline(BM_ENTER_BUSINESS_TYPE_MESSAGE);
         if (!scanner.hasNextLine()) {
-            //System.err.println("Error: No input detected. Exiting...");
             UI.printErrorMessage(BM_NO_INPUT_ERROR_MESSAGE);
             return;
         }
@@ -176,7 +171,6 @@ public class BusynessManager {
         businessID = id;
         credentials.put(businessID, businessPassword);
 
-        //System.out.println("Business setup complete!");
         UI.printMessage(BM_SETUP_COMPLETE_MESSAGE);
     }
 
@@ -197,14 +191,14 @@ public class BusynessManager {
      * @param scanner The Scanner object for user input.
      */
     public void run(Scanner scanner) {
-        //System.out.println("Busyness Manager is ready. Type 'help' for commands.");
         UI.printMessage(BM_READY_MESSAGE);
+
         while (true) {
-            //System.out.print(">");
             UI.printMessageWithoutNewline(BM_WAITING_INPUT_MESSAGE);
+
             String input = scanner.nextLine();
+
             if (input.equalsIgnoreCase(BM_EXIT_KEYWORD)) {
-                //System.out.println("Exiting Busyness Manager...");
                 UI.printMessage(BM_EXIT_MESSAGE);
                 break;
             }
