+++ src/main/java/busynessmanager/exceptions/InvalidStringException.java
@@ -4,7 +4,6 @@ package busynessmanager.exceptions;
  * Exception for an invalid String.
  */
 public class InvalidStringException extends Exception {
-
     /**
      * Constructs the InvalidString exception.
      */
