+++ src/main/java/busynessmanager/managers/SalesManager.java
@@ -1,11 +1,12 @@
 package busynessmanager.managers;
 
-import busynessmanager.UI_Constants.UI;
-//import static busynessmanager.UI_Constants.Constants.*; // This will fail style check
-import static busynessmanager.UI_Constants.Constants.NEWLINE;
-import static busynessmanager.UI_Constants.Constants.SM_MINIMUM_QTY_SOLD_MESSAGE;
-import static busynessmanager.UI_Constants.Constants.SM_CLEARED_FORMAT;
-import static busynessmanager.UI_Constants.Constants.SM_RECORDED_FORMAT;
+import busynessmanager.ui.UI;
+
+import static busynessmanager.constants.Constants.NEWLINE;
+import static busynessmanager.constants.Constants.SM_MINIMUM_QTY_SOLD_MESSAGE;
+import static busynessmanager.constants.Constants.SM_CLEARED_FORMAT;
+import static busynessmanager.constants.Constants.SM_RECORDED_FORMAT;
+
 
 /**
  * Updates sales transactions, decrements inventory and increases quantity sold
@@ -13,42 +14,47 @@ import static busynessmanager.UI_Constants.Constants.SM_RECORDED_FORMAT;
 public class SalesManager {
     private final InventoryManager inventory;
 
+
+    /**
+     * Constructs the SalesManager class using an existing InventoryManager.
+     *
+     * @param inventory Existing InventoryManager for the business.
+     */
     public SalesManager(InventoryManager inventory) {
         this.inventory = inventory;
     }
 
+
     /**
-     * Record a sale of a quantity of a product by its ID
+     * Records a sale of a quantity of a product by its ID.
      *
-     * @param id Product ID
-     * @param qtySold Quantity of product sold
+     * @param id Product ID.
+     * @param qtySold Quantity of product sold.
      */
     public void recordSale(String id, int qtySold) {
         if (qtySold <= 0) {
-            //System.out.println("Quantity sold must be greater than 0.");
             UI.printMessage(SM_MINIMUM_QTY_SOLD_MESSAGE);
             return;
         }
-        inventory.updateProductQuantity(id, qtySold);
 
-        //System.out.println("Sale recorded: Product ID " + id + ", Quantity Sold: " + qtySold);
+        inventory.updateProductQuantity(id, qtySold);
         UI.printFormattedMessage(SM_RECORDED_FORMAT + NEWLINE, id, qtySold);
     }
 
     /**
-     * Clear sales for a product by its ID
+     * Clears sales for a product by its ID.
      *
-     * @param id product ID
+     * @param id product ID.
      */
     public void clearSales(String id) {
         inventory.resetProductSales(id);
-        //System.out.println("Sales cleared: Product ID " + id);
         UI.printFormattedMessage(SM_CLEARED_FORMAT + NEWLINE, id);
     }
 
     /**
-     * required for SY's RevenueCalculator to function
-     * @return the inventory
+     * Returns the InventoryManager used in the SalesManager class.
+     *
+     * @return the InventoryManager used by the business.
      */
     public InventoryManager getInventory() {
         return inventory;
