+++ src/main/java/busynessmanager/parser/CommandParser.java
@@ -1,51 +1,51 @@
 package busynessmanager.parser;
 
-import busynessmanager.UI_Constants.UI;
 import busynessmanager.managers.InventoryManager;
 import busynessmanager.managers.SalesManager;
 import busynessmanager.managers.SearchManager;
 import busynessmanager.revenue.RevenueCalculator;
+import busynessmanager.ui.UI;
 
-import static busynessmanager.UI_Constants.Constants.WHITESPACE;
-import static busynessmanager.UI_Constants.Constants.EMPTY_STRING;
-import static busynessmanager.UI_Constants.Constants.INDEX_0;
-import static busynessmanager.UI_Constants.Constants.INDEX_1;
-import static busynessmanager.UI_Constants.Constants.INDEX_2;
-import static busynessmanager.UI_Constants.Constants.INDEX_3;
-import static busynessmanager.UI_Constants.Constants.INDEX_4;
-import static busynessmanager.UI_Constants.Constants.INDEX_5;
-import static busynessmanager.UI_Constants.Constants.INDEX_6;
-import static busynessmanager.UI_Constants.Constants.INDEX_7;
-import static busynessmanager.UI_Constants.Constants.CP_NAME;
-import static busynessmanager.UI_Constants.Constants.CP_COMMAND_SEPARATOR_INDEX;
-import static busynessmanager.UI_Constants.Constants.CP_ADD_COMMAND;
-import static busynessmanager.UI_Constants.Constants.CP_DELETE_COMMAND;
-import static busynessmanager.UI_Constants.Constants.CP_UPDATE_COMMAND;
-import static busynessmanager.UI_Constants.Constants.CP_PRINT_COMMAND;
-import static busynessmanager.UI_Constants.Constants.CP_SOLD_COMMAND;
-import static busynessmanager.UI_Constants.Constants.CP_CLEAR_COMMAND;
-import static busynessmanager.UI_Constants.Constants.CP_REVENUE_COMMAND;
-import static busynessmanager.UI_Constants.Constants.CP_SEARCH_COMMAND;
-import static busynessmanager.UI_Constants.Constants.CP_ID_REGEX;
-import static busynessmanager.UI_Constants.Constants.CP_SPLIT_REGEX;
-import static busynessmanager.UI_Constants.Constants.CP_ID_FLAG;
-import static busynessmanager.UI_Constants.Constants.CP_NAME_FLAG;
-import static busynessmanager.UI_Constants.Constants.CP_PRICE_FLAG;
-import static busynessmanager.UI_Constants.Constants.CP_QUANTITY_FLAG;
-import static busynessmanager.UI_Constants.Constants.CP_INVALID_COMMAND_MESSAGE;
-import static busynessmanager.UI_Constants.Constants.CP_INVALID_FLAG_MESSAGE;
-import static busynessmanager.UI_Constants.Constants.CP_INVALID_FLAG_MESSAGE_2;
-import static busynessmanager.UI_Constants.Constants.CP_INVALID_FLAG_MESSAGE_3;
-import static busynessmanager.UI_Constants.Constants.CP_INVALID_FLAG_MESSAGE_4;
-import static busynessmanager.UI_Constants.Constants.CP_INVALID_FLAG_MESSAGE_5;
-import static busynessmanager.UI_Constants.Constants.CP_INVALID_FLAG_MESSAGE_6;
-import static busynessmanager.UI_Constants.Constants.CP_INVALID_ID_MESSAGE;
-import static busynessmanager.UI_Constants.Constants.CP_INVALID_ID_FORMAT_MESSAGE;
-import static busynessmanager.UI_Constants.Constants.CP_INVALID_NUMERAL_MESSAGE;
-import static busynessmanager.UI_Constants.Constants.CP_INVALID_NUMERAL_MESSAGE_2;
-import static busynessmanager.UI_Constants.Constants.CP_NEGATIVE_QUANTITY_MESSAGE;
-import static busynessmanager.UI_Constants.Constants.CP_NEGATIVE_PRICE_MESSAGE;
-import static busynessmanager.UI_Constants.Constants.CP_EXCEPTION_LOG_MESSAGE;
+import static busynessmanager.constants.Constants.WHITESPACE;
+import static busynessmanager.constants.Constants.EMPTY_STRING;
+import static busynessmanager.constants.Constants.INDEX_0;
+import static busynessmanager.constants.Constants.INDEX_1;
+import static busynessmanager.constants.Constants.INDEX_2;
+import static busynessmanager.constants.Constants.INDEX_3;
+import static busynessmanager.constants.Constants.INDEX_4;
+import static busynessmanager.constants.Constants.INDEX_5;
+import static busynessmanager.constants.Constants.INDEX_6;
+import static busynessmanager.constants.Constants.INDEX_7;
+import static busynessmanager.constants.Constants.CP_NAME;
+import static busynessmanager.constants.Constants.CP_COMMAND_SEPARATOR_INDEX;
+import static busynessmanager.constants.Constants.CP_ADD_COMMAND;
+import static busynessmanager.constants.Constants.CP_DELETE_COMMAND;
+import static busynessmanager.constants.Constants.CP_UPDATE_COMMAND;
+import static busynessmanager.constants.Constants.CP_PRINT_COMMAND;
+import static busynessmanager.constants.Constants.CP_SOLD_COMMAND;
+import static busynessmanager.constants.Constants.CP_CLEAR_COMMAND;
+import static busynessmanager.constants.Constants.CP_REVENUE_COMMAND;
+import static busynessmanager.constants.Constants.CP_SEARCH_COMMAND;
+import static busynessmanager.constants.Constants.CP_ID_REGEX;
+import static busynessmanager.constants.Constants.CP_SPLIT_REGEX;
+import static busynessmanager.constants.Constants.CP_ID_FLAG;
+import static busynessmanager.constants.Constants.CP_NAME_FLAG;
+import static busynessmanager.constants.Constants.CP_PRICE_FLAG;
+import static busynessmanager.constants.Constants.CP_QUANTITY_FLAG;
+import static busynessmanager.constants.Constants.CP_INVALID_COMMAND_MESSAGE;
+import static busynessmanager.constants.Constants.CP_INVALID_FLAG_MESSAGE;
+import static busynessmanager.constants.Constants.CP_INVALID_FLAG_MESSAGE_2;
+import static busynessmanager.constants.Constants.CP_INVALID_FLAG_MESSAGE_3;
+import static busynessmanager.constants.Constants.CP_INVALID_FLAG_MESSAGE_4;
+import static busynessmanager.constants.Constants.CP_INVALID_FLAG_MESSAGE_5;
+import static busynessmanager.constants.Constants.CP_INVALID_FLAG_MESSAGE_6;
+import static busynessmanager.constants.Constants.CP_INVALID_ID_MESSAGE;
+import static busynessmanager.constants.Constants.CP_INVALID_ID_FORMAT_MESSAGE;
+import static busynessmanager.constants.Constants.CP_INVALID_NUMERAL_MESSAGE;
+import static busynessmanager.constants.Constants.CP_INVALID_NUMERAL_MESSAGE_2;
+import static busynessmanager.constants.Constants.CP_NEGATIVE_QUANTITY_MESSAGE;
+import static busynessmanager.constants.Constants.CP_NEGATIVE_PRICE_MESSAGE;
+import static busynessmanager.constants.Constants.CP_EXCEPTION_LOG_MESSAGE;
 
 import busynessmanager.exceptions.InvalidStringException;
 import busynessmanager.exceptions.InvalidCommandException;
@@ -54,12 +54,11 @@ import busynessmanager.exceptions.NumberParsingFailedException;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
+
 /**
  * Class for parsing the user input, and executing the appropriate methods.
  */
 public class CommandParser {
-
-
     private static final Logger logger = Logger.getLogger(CP_NAME);
     private final InventoryManager inventoryManager;
     private final SalesManager salesManager;
@@ -78,7 +77,6 @@ public class CommandParser {
         this.searchManager = new SearchManager(inventoryManager);
     }
 
-
     /**
      * Constructs the CommandParser class given existing systems for the business.
      *
@@ -87,7 +85,6 @@ public class CommandParser {
      * @param revenueCalculator existing RevenueCalculator for the business.
      * @param searchManager existing SearchManager for the business.
      */
-
     public CommandParser(InventoryManager inventoryManager, SalesManager salesManager,
                          RevenueCalculator revenueCalculator, SearchManager searchManager) {
         this.inventoryManager = inventoryManager;
@@ -95,6 +92,7 @@ public class CommandParser {
         this.revenueCalculator = revenueCalculator;
         this.searchManager = searchManager;
     }
+
 
     /**
      * Parses the user input for execution.
