+++ src/main/java/busynessmanager/revenue/RevenueCalculator.java
@@ -1,57 +1,62 @@
 package busynessmanager.revenue;
 
-import busynessmanager.UI_Constants.UI;
+import busynessmanager.product.Product;
 import busynessmanager.managers.InventoryManager;
 import busynessmanager.managers.SalesManager;
-import busynessmanager.product.Product;
-import static busynessmanager.UI_Constants.Constants.MINIMUM_VALUE;
-import static busynessmanager.UI_Constants.Constants.NEWLINE;
-import static busynessmanager.UI_Constants.Constants.RC_INDIVIDUAL_REVENUE_FORMAT;
-import static busynessmanager.UI_Constants.Constants.RC_TOTAL_REVENUE_FORMAT;
+import busynessmanager.ui.UI;
+
+import static busynessmanager.constants.Constants.MINIMUM_VALUE;
+import static busynessmanager.constants.Constants.NEWLINE;
+import static busynessmanager.constants.Constants.RC_INDIVIDUAL_REVENUE_FORMAT;
+import static busynessmanager.constants.Constants.RC_TOTAL_REVENUE_FORMAT;
 
 import java.util.HashMap;
 
+
 /**
- * RevenueCalculator is used to calculate the various revenues that is produced in the Busyness
+ * RevenueCalculator is used to calculate the various revenues that is produced in BusynessManager.
  */
 public class RevenueCalculator {
     private final SalesManager sm;
 
+
     /**
-     * Constructor
-     * @param sm Pre-existing SalesManager to be used for instantiation
+     * Constructs the RevenueCalculator class using an existing SalesManager.
+     *
+     * @param sm Pre-existing SalesManager to be used for instantiation.
      */
     public RevenueCalculator(SalesManager sm) {
         this.sm = sm;
     }
 
+
     /**
-     * Sums up the total revenue of all Products in the SalesManager (which has an InventoryManager)
-     * @return total revenue at this point of time
+     * Sums up the total revenue of all Products in the SalesManager (which has an InventoryManager).
      */
-    public double computeTotalRevenue() {
+    public void computeTotalRevenue() {
         InventoryManager currentInventory = sm.getInventory();
         HashMap<String, Product> currentProductList = currentInventory.returnProductList();
 
         double totalRevenue = MINIMUM_VALUE;
+
         for (Product product : currentProductList.values()) {
             double individualRevenue = product.getPrice() * product.getQuantitySold();
+
             UI.printFormattedMessage(RC_INDIVIDUAL_REVENUE_FORMAT + NEWLINE,
-                product.getName(), individualRevenue);
+                    product.getName(), individualRevenue);
+
             totalRevenue += individualRevenue;
         }
 
         UI.printFormattedMessage(RC_TOTAL_REVENUE_FORMAT + NEWLINE, totalRevenue);
-
-        return totalRevenue;
     }
 
     /**
-     * Returns the revenue of a given product
-     * @param id The String ID of the product to be queried
-     * @return Revenue from selling the given product
+     * Returns the revenue of a given product.
+     *
+     * @param id The String ID of the product to be queried.
      */
-    public double computeProductRevenue(String id) {
+    public void computeProductRevenue(String id) {
         InventoryManager currentInventory = sm.getInventory();
         HashMap<String, Product> currentProductList = currentInventory.returnProductList();
 
@@ -59,9 +64,8 @@ public class RevenueCalculator {
         Product product = currentProductList.get(id);
 
         double individualRevenue = product.getPrice() * product.getQuantitySold();
+
         UI.printFormattedMessage(RC_INDIVIDUAL_REVENUE_FORMAT + NEWLINE,
             product.getName(), individualRevenue);
-
-        return individualRevenue;
     }
 }
